+++ common-test/build.gradle.kts
@@ -32,4 +32,5 @@ dependencies {
     implementation(libs.kotlinx.coroutines.test)
     implementation(libs.hilt.android.testing)
     implementation(libs.truth)
+    implementation(libs.androidx.paging.runtime)
 }
